Database Singleton, Validate Received Json
+ Add and delete methods with json validation.

+++ DataRepresentationBackend/src/main/java/DataRepresentationBackend/Controllers/HTTPController.java
@@ -24,35 +24,42 @@ public class HTTPController {
     public ResponseEntity<Test> getTest() {
         Test response = new Test();
         response.dummyNumber = databaseService.test();
-        return new ResponseEntity<Test>(response, HttpStatus.OK);
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
     @ResponseBody
     public ResponseEntity<Message> addTemporaryData(@RequestBody TemporaryData data) {
         try {
+            if (data.getElementType() == null || data.getFloor() == null || data.getIsExitWay() == null
+                    || data.getRoom() == null || data.getX1() == null || data.getX2() == null ||
+                    data.getY1() == null || data.getY2() == null) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             databaseService.addData(data);
             databaseService.commit();
-            return new ResponseEntity<Message>(new Message("Insert operation success!"), HttpStatus.OK);
-        }
-        catch (Exception ex) {
-            return new ResponseEntity<Message>(new Message("Insert operation failed!"), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(new Message("Insert operation success!"), HttpStatus.OK);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Message("Insert operation failed!"), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @RequestMapping(value = "/delete", method = RequestMethod.POST)
+    @RequestMapping(value = "/delete", method = RequestMethod.DELETE)
     @ResponseBody
     public ResponseEntity<Message> deleteTemporaryData(@RequestBody TemporaryData data) {
         try {
+            if (data.getElementType() == null || data.getFloor() == null || data.getIsExitWay() == null
+                    || data.getRoom() == null || data.getX1() == null || data.getX2() == null ||
+                    data.getY1() == null || data.getY2() == null) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             databaseService.deleteData(data);
             databaseService.commit();
-            return new ResponseEntity<Message>(new Message("Delete operation success!"), HttpStatus.OK);
-        }
-        catch (OperationNotSupportedException ex) {
-            return new ResponseEntity<Message>(new Message(ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-        catch (Exception ex) {
-            return new ResponseEntity<Message>(new Message("Delete operation failed!"), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(new Message("Delete operation success!"), HttpStatus.OK);
+        } catch (OperationNotSupportedException ex) {
+            return new ResponseEntity<>(new Message(ex.getMessage()), HttpStatus.BAD_REQUEST);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Message("Delete operation failed!"), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
